requirements.txt
streamlit
psycopg2-binary
sentence-transformers
pandas
python-dotenv
pymongo